move send the schedulizer button
